backend/requirements.txt
2captcha-python==1.1.3
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cohere==3.2.3
cryptography==39.0.0
defusedxml==0.7.1
dj-database-url==1.2.0
Django==3.2
django-allauth==0.52.0
django-cors-headers==3.13.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
oauthlib==3.2.2
outcome==1.2.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
PySocks==1.7.1
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
revChatGPT==0.3.1
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tls-client==0.1.8
trio==0.22.0
trio-websocket==0.9.2
undetected-chromedriver==3.2.1
urllib3==1.26.14
websockets==10.4
wsproto==1.2.0
